+++ edsnlp/pipes/misc/measurements/__init__.py
@@ -1,4 +1,0 @@
-from edsnlp.pipes.misc.measurements.measurements import MeasurementsMatcher
-from edsnlp.pipes.misc.measurements.patterns import *
-
-from . import factory